LP-1919 refactor psc validation
