+++ src/config/constants.ts
@@ -175,3 +175,11 @@ export const VALID_CHARACTERS_REGEX =
 export const LEGAL_ENTITY_NAME_FIELD = "legal_entity_name";
 export const LEGAL_FORM_FIELD = "legal_form";
 export const GOVERNING_LAW_FIELD = "governing_law";
+export const TITLE_FIELD = "title";
+export const FORENAME_FIELD = "forename";
+export const MIDDLE_NAMES_FIELD = "middle_names";
+export const SURNAME_FIELD = "surname";
+export const DATE_OF_BIRTH_FIELD = "date_of_birth";
+export const NATIONALITY1_FIELD = "nationality1";
+export const NATIONALITY2_FIELD = "nationality2";
+export const CONSENT_CHECKED_FIELD = "consent_checked";
